update torch calls and use CPU instead of CUDA

+++ misc/atlanta_transform.py
@@ -25,16 +25,16 @@ class A2P(nn.Module):
         self.gpu = gpu
 
         R_lst = []
-        theta_lst = np.array([-90, 0, 90, 180], np.float) / 180 * np.pi
-        phi_lst = np.array([90, -90], np.float) / 180 * np.pi
+        theta_lst = np.array([-90, 0, 90, 180], float) / 180 * np.pi
+        phi_lst = np.array([90, -90], float) / 180 * np.pi
 
         for theta in theta_lst:
-            angle_axis = theta * np.array([0, 1, 0], np.float)
+            angle_axis = theta * np.array([0, 1, 0], float)
             R = cv2.Rodrigues(angle_axis)[0]
             R_lst.append(R)
 
         for phi in phi_lst:
-            angle_axis = phi * np.array([1, 0, 0], np.float)
+            angle_axis = phi * np.array([1, 0, 0], float)
             R = cv2.Rodrigues(angle_axis)[0]
             R_lst.append(R)
 
@@ -63,11 +63,11 @@ class A2P(nn.Module):
 
         interval = w_len / (out_dim - 1)
         
-        z_map = np.zeros([out_dim, out_dim], np.float32) + radius
+        z_map = np.zeros([out_dim, out_dim], float) + radius
         x_map = np.tile((np.arange(out_dim) - c_x) * interval, [out_dim, 1])
         y_map = np.tile((np.arange(out_dim) - c_y) * interval, [out_dim, 1]).T
         D = np.sqrt(x_map**2 + y_map**2 + z_map**2)
-        xyz = np.zeros([out_dim, out_dim, 3], np.float)
+        xyz = np.zeros([out_dim, out_dim, 3], float)
         xyz[:, :, 0] = (radius / D) * x_map[:, :]
         xyz[:, :, 1] = (radius / D) * y_map[:, :]
         xyz[:, :, 2] = (radius / D) * z_map[:, :]
